+++ src/ios/WikitudeSDK.h
@@ -2,10 +2,12 @@
 //  WikitudeSDK.h
 //  WikitudeSDK
 //
-//  Copyright (c) 2014 Wikitude. All rights reserved.
+//  Copyright (c) 2017 Wikitude. All rights reserved.
 //
 
 #import "WTArchitectView.h"
 #import "WTWikitudeTypes.h"
 #import "WTNavigation.h"
-#import "WTStartupConfiguration.h"
+#import "WTArchitectStartupConfiguration.h"
+#import "WTSDKBuildInformation.h"
+#import "WTAuthorizationRequestManager.h"